+++ packages/app/src/scenes/collaboration/stores/SpaceStore_OLD/index.ts
@@ -1,0 +1,1 @@
+export * from './SpaceStore_OLD';
